// apps/web/react/app/src/generated/api/models/Column.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Column = {
    /**
     * A URL to the JSON Schema for this object.
     */
    readonly $schema?: string;
    boardId: string;
    createdAt: string;
    id: string;
    position: number;
    title: string;
    updatedAt: string;
};